Remove links for titles in page headers
Currently each page can have `@title_url` variable, which is used as the
destination of the page header's title link. But currently the variable
isn't used anywhere, and the title is a link to itself for each page as
a result. It doesn't seem the initial intention and some pages (e.g.
POST /multifactor_auth) generate a link to nothing.
Therefore I think it might be better if such title links would be
removed and let them just strings.